クラス名を修正し、MoveをLiteralMove に BoardMoveをMove に変更しました。
