+++ Ragnarok.Presentation.Shogi/View/ShogiUIElement3D.move.cs
@@ -309,7 +309,7 @@ namespace Ragnarok.Presentation.Shogi.View
         /// <summary>
         /// 指し手が実際に着手可能か確認します。
         /// </summary>
-        private bool CanMove(BoardMove move, MoveFlags flags = MoveFlags.DoMoveDefault)
+        private bool CanMove(Move move, MoveFlags flags = MoveFlags.DoMoveDefault)
         {
             var tmp = Board.Clone();
 
@@ -366,12 +366,12 @@ namespace Ragnarok.Presentation.Shogi.View
         {
             var srcSquare = this.movingPiece.Square;
             var piece = this.movingPiece.Piece;
-            BoardMove move = null;
+            Move move = null;
 
             if (srcSquare != null)
             {
                 // 駒の移動の場合
-                move = BoardMove.CreateMove(
+                move = Move.CreateMove(
                     piece.BWType, srcSquare, dstSquare, piece.Piece, false);
 
                 // 成／不成りのダイアログを出す前に着手可能か確認します。
@@ -396,7 +396,7 @@ namespace Ragnarok.Presentation.Shogi.View
             else
             {
                 // 駒打ちの場合
-                move = BoardMove.CreateDrop(
+                move = Move.CreateDrop(
                     piece.BWType, dstSquare, piece.PieceType);
 
                 if (!CanMove(move))
@@ -413,7 +413,7 @@ namespace Ragnarok.Presentation.Shogi.View
         /// <summary>
         /// 実際に指し手を進めます。
         /// </summary>
-        private void DoMove(BoardMove move)
+        private void DoMove(Move move)
         {
             if (move == null || !move.Validate())
             {
