+++ Ragnarok.Shogi/Move.cs
@@ -4,209 +4,46 @@ using System.Linq;
 using System.Text;
 using System.Runtime.Serialization;
 
+using Ragnarok.Utility;
+
 namespace Ragnarok.Shogi
 {
     /// <summary>
-    /// 駒の段方向の動き方の種類です。
+    /// 盤面上での差し手を示します。
     /// </summary>
-    [DataContract()]
-    public enum RankMoveType
-    {
-        /// <summary>
-        /// 特に無し
-        /// </summary>
-        [EnumMember()]
-        None = 0,
-        /// <summary>
-        /// 上がる
-        /// </summary>
-        [EnumMember()]
-        Up = 1,
-        /// <summary>
-        /// 引く
-        /// </summary>
-        [EnumMember()]
-        Back = 2,
-        /// <summary>
-        /// 寄る
-        /// </summary>
-        [EnumMember()]
-        Sideways = 3,
-    }
-
-    /// <summary>
-    /// 駒の列方向の相対位置の種類です。
-    /// </summary>
-    [DataContract()]
-    public enum RelFileType
-    {
-        /// <summary>
-        /// 特になし。
-        /// </summary>
-        [EnumMember()]
-        None = 0,
-        /// <summary>
-        /// 右
-        /// </summary>
-        [EnumMember()]
-        Right = 1,
-        /// <summary>
-        /// 左
-        /// </summary>
-        [EnumMember()]
-        Left = 2,
-        /// <summary>
-        /// 直
-        /// </summary>
-        [EnumMember()]
-        Straight = 3,
-    }
-
-    /// <summary>
-    /// 駒打ちなどのアクションを記述します。
-    /// </summary>
-    [DataContract()]
-    public enum ActionType
-    {
-        /// <summary>
-        /// 特になし。
-        /// </summary>
-        [EnumMember()]
-        None = 0,
-        /// <summary>
-        /// 成り
-        /// </summary>
-        [EnumMember()]
-        Promote = 1,
-        /// <summary>
-        /// 不成
-        /// </summary>
-        [EnumMember()]
-        Unpromote = 2,
-        /// <summary>
-        /// 打ち
-        /// </summary>
-        [EnumMember()]
-        Drop = 3,
-    }
-
-    /// <summary>
-    /// 将棋の各指し手を保持します。
-    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// シリアライズについて
+    /// 
+    /// このクラスは<see cref="Board"/>クラスと一緒にシリアライズされますが、
+    /// 200手,300手の局面になってくると、シリアライズ後のデータ量が
+    /// 3kbや4kbになってしまいます。
+    /// そのためこのクラスでは特別に、シリアライズ時にデータの
+    /// 圧縮を行っています。
+    /// </para>
+    /// </remarks>
     [DataContract()]
     public class Move : IEquatable<Move>
     {
         /// <summary>
-        /// 元の指し手を取得または設定します。
+        /// オブジェクトのコピーを作成します。
         /// </summary>
-        [DataMember(Order = 1, IsRequired = true)]
-        public string OriginalText
+        public Move Clone()
         {
-            get;
-            set;
+            return (Move)MemberwiseClone();
         }
 
         /// <summary>
-        /// 同○○かどうかを取得または設定します。
+        /// 特殊な指し手の場合はその種類を取得または設定します。
         /// </summary>
-        [DataMember(Order = 2, IsRequired = true)]
-        public bool SameAsOld
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 駒の到達地点の筋を取得または設定します。
-        /// </summary>
-        [DataMember(Order = 3, IsRequired = true)]
-        public int File
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 駒の到達地点の段を取得または設定します。
-        /// </summary>
-        [DataMember(Order = 4, IsRequired = true)]
-        public int Rank
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 駒の到達位置を取得または設定します。
-        /// </summary>
-        public Square DstSquare
-        {
-            get
-            {
-                return new Square(File, Rank);
-            }
-            set
-            {
-                if ((object)value != null)
-                {
-                    File = value.File;
-                    Rank = value.Rank;
-                }
-            }
-        }
-
-        /// <summary>
-        /// 駒の種類を取得または設定します。
-        /// </summary>
-        [DataMember(Order = 5, IsRequired = true)]
-        public Piece Piece
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 移動前の駒の相対位置を取得または設定します。
-        /// </summary>
-        [DataMember(Order = 6, IsRequired = true)]
-        public RelFileType RelFileType
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 駒の動き方の種類を取得または設定します。
-        /// </summary>
-        [DataMember(Order = 7, IsRequired = true)]
-        public RankMoveType RankMoveType
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 駒打ちなどのアクションを取得または設定します。
-        /// </summary>
-        [DataMember(Order = 8, IsRequired = true)]
-        public ActionType ActionType
-        {
-            get;
-            set;
-        }
-
-        /// <summary>
-        /// 投了などの特殊な指し手を取得または設定します。
-        /// </summary>
-        [DataMember(Order = 9, IsRequired = true)]
         public SpecialMoveType SpecialMoveType
         {
             get;
-            set;
+            private set;
         }
 
         /// <summary>
-        /// 特殊な指し手であるか調べます。
+        /// 特殊な指し手であるかどうかを取得します。
         /// </summary>
         public bool IsSpecialMove
         {
@@ -214,115 +51,217 @@ namespace Ragnarok.Shogi
         }
 
         /// <summary>
-        /// 手番を取得または設定します。
+        /// 駒を移動する場合の対象となる駒を取得または設定します。
         /// </summary>
-        [DataMember(Order = 10, IsRequired = true)]
-        public BWType BWType
+        public Piece MovePiece
         {
             get;
-            set; 
+            private set;
         }
 
         /// <summary>
-        /// (もしあれば)移動前の情報を取得または設定します。
+        /// 先手の手か後手の手かを取得または設定します。
         /// </summary>
-        [DataMember(Order = 11, IsRequired = false)]
+        public BWType BWType
+        {
+            get;
+            private set;
+        }
+        
+        /// <summary>
+        /// 駒の移動先を取得または設定します。
+        /// </summary>
+        public Square DstSquare
+        {
+            get;
+            private set;
+        }
+
+        /// <summary>
+        /// 駒の移動元を取得または設定します。
+        /// </summary>
+        /// <remarks>
+        /// 駒打ちの場合はnullになります。
+        /// </remarks>
         public Square SrcSquare
+        {
+            get;
+            private set;
+        }
+
+        /// <summary>
+        /// 駒を成るかどうかを取得または設定します。
+        /// </summary>
+        public bool IsPromote
         {
             get;
             set;
         }
 
         /// <summary>
-        /// オブジェクトの簡易コピーを作成します。
+        /// 駒を打つ場合の駒を取得または設定します。
         /// </summary>
-        public Move Clone()
+        public PieceType DropPieceType
         {
-            return (Move)this.MemberwiseClone();
+            get;
+            private set;
         }
 
         /// <summary>
-        /// オブジェクトの各プロパティが正しく設定されているか調べます。
+        /// 駒打ち・成りなどの動作を取得します。
+        /// </summary>
+        public ActionType ActionType
+        {
+            get
+            {
+                if (IsSpecialMove)
+                {
+                    return ActionType.None;
+                }
+                else if (DropPieceType != PieceType.None)
+                {
+                    return ActionType.Drop;
+                }
+                else if (SrcSquare != null && MovePiece != null)
+                {
+                    return (IsPromote ? ActionType.Promote : ActionType.None);
+                }
+
+                throw new ShogiException(
+                    "指し手が正しくありません。");
+            }
+        }
+
+        /// <summary>
+        /// 取った駒を取得または設定します。
+        /// </summary>
+        /// <remarks>
+        /// 戻る操作のために必要です。
+        /// </remarks>
+        public Piece TookPiece
+        {
+            get;
+            set;
+        }
+
+        /// <summary>
+        /// 一手前の指し手と着手位置が同じかどうかを取得または設定します。
+        /// </summary>
+        /// <remarks>
+        /// 「同飛」などと出力するためのもので、それ以上の意味はありません。
+        /// </remarks>
+        public bool HasSameSquareAsPrev
+        {
+            get;
+            set;
+        }
+
+        /// <summary>
+        /// 指し手を文字列化します。
+        /// </summary>
+        public override string ToString()
+        {
+            if (SpecialMoveType != SpecialMoveType.None)
+            {
+                return string.Format(
+                    "{0}{1}",
+                    Stringizer.ToString(BWType),
+                    EnumEx.GetLabel(SpecialMoveType));
+            }
+            else if (ActionType == ActionType.Drop)
+            {
+                return string.Format(
+                    "{0}{1}{2}{3}打",
+                    Stringizer.ToString(BWType),
+                    IntConverter.Convert(NumberType.Big, DstSquare.File),
+                    IntConverter.Convert(NumberType.Kanji, DstSquare.Rank),
+                    Stringizer.ToString(DropPieceType));
+            }
+            else if (HasSameSquareAsPrev)
+            {
+                return string.Format(
+                    "{0}同　{1}{2}({3}{4})",
+                    Stringizer.ToString(BWType),
+                    MovePiece,
+                    Stringizer.ToString(ActionType),
+                    SrcSquare.File,
+                    SrcSquare.Rank);
+            }
+            else
+            {
+                return string.Format(
+                    "{0}{1}{2}{3}{4}({5}{6})",
+                    Stringizer.ToString(BWType),
+                    IntConverter.Convert(NumberType.Big, DstSquare.File),
+                    IntConverter.Convert(NumberType.Kanji, DstSquare.Rank),
+                    MovePiece,
+                    Stringizer.ToString(ActionType),
+                    SrcSquare.File,
+                    SrcSquare.Rank);
+            }
+        }
+
+        /// <summary>
+        /// オブジェクトの妥当性を検証します。
         /// </summary>
         public bool Validate()
         {
-            if (IsSpecialMove)
+            if (BWType == BWType.None)
             {
-                return true;
+                return false;
             }
 
-            if (!SameAsOld)
+            if (SpecialMoveType != SpecialMoveType.None)
             {
-                if (File < 1 || 9 < File)
+                // 特に確認は行いません。
+            }
+            else if (ActionType == ActionType.Drop)
+            {
+                // 駒打ちの場合
+                if (DstSquare == null || !DstSquare.Validate())
+                {
+                    return false;
+                }
+                
+                if (SrcSquare != null)
                 {
                     return false;
                 }
 
-                if (Rank < 1 || 9 < Rank)
+                if (MovePiece != null)
+                {
+                    return false;
+                }
+
+                if (IsPromote)
                 {
                     return false;
                 }
             }
-
-            if (Piece == null || !Piece.Validate())
+            else
             {
-                return false;
-            }
+                // 駒打ちでない場合
+                if (DstSquare == null || !DstSquare.Validate())
+                {
+                    return false;
+                }
 
-            if (!Enum.IsDefined(typeof(RelFileType), RelFileType))
-            {
-                return false;
-            }
+                if (SrcSquare == null || !SrcSquare.Validate())
+                {
+                    return false;
+                }
 
-            if (!Enum.IsDefined(typeof(RankMoveType), RankMoveType))
-            {
-                return false;
-            }
-
-            if (!Enum.IsDefined(typeof(ActionType), ActionType))
-            {
-                return false;
-            }
-
-            // 竜打つ、同銀打ち、銀上打つ、などは存在しません。
-            if (ActionType == ActionType.Drop &&
-                (Piece.IsPromoted ||
-                 SameAsOld ||
-                 RelFileType != RelFileType.None ||
-                 RankMoveType != RankMoveType.None))
-            {
-                return false;
-            }
-
-            // 金と玉、成り駒は成れません。
-            if ((ActionType == ActionType.Promote ||
-                 ActionType == ActionType.Unpromote) &&
-                (Piece.IsPromoted ||
-                 Piece.PieceType == PieceType.Kin ||
-                 Piece.PieceType == PieceType.Gyoku))
-            {
-                return false;
-            }
-
-            // 移動前の位置は、nullでなければ内容を調べます。
-            if (SrcSquare != null && !SrcSquare.Validate())
-            {
-                return false;
+                if (MovePiece == null)
+                {
+                    return false;
+                }
             }
 
             return true;
         }
 
         /// <summary>
-        /// 指し手を文字列に変換します。
-        /// </summary>
-        public override string ToString()
-        {
-            return Stringizer.ToString(this);
-        }
-
-        /// <summary>
-        /// オブジェクトの比較を行います。
+        /// オブジェクトの等値性を調べます。
         /// </summary>
         public override bool Equals(object obj)
         {
@@ -336,16 +275,11 @@ namespace Ragnarok.Shogi
         }
 
         /// <summary>
-        /// オブジェクトの比較を行います。
+        /// オブジェクトの等値性を調べます。
         /// </summary>
         public bool Equals(Move other)
         {
             if ((object)other == null)
-            {
-                return false;
-            }
-
-            if (SpecialMoveType != other.SpecialMoveType)
             {
                 return false;
             }
@@ -355,28 +289,32 @@ namespace Ragnarok.Shogi
                 return false;
             }
 
-            if (SameAsOld != other.SameAsOld)
+            if (DstSquare != other.DstSquare)
             {
                 return false;
             }
 
-            // 指し手が同○○でないときだけ、段・列の判定をします。
-            if (!SameAsOld &&
-                (File != other.File || Rank != other.Rank))
-            {
-                return false;
-            }
-
-            if (RelFileType != other.RelFileType ||
-                RankMoveType != other.RankMoveType ||
-                ActionType != other.ActionType ||
-                Piece != other.Piece)
-            {
-                return false;
-            }
-
-            // 移動前の位置は判定に使うべきかどうか迷う。
             if (SrcSquare != other.SrcSquare)
+            {
+                return false;
+            }
+
+            if (IsPromote != other.IsPromote)
+            {
+                return false;
+            }
+
+            if (MovePiece != other.MovePiece)
+            {
+                return false;
+            }
+
+            if (DropPieceType != other.DropPieceType)
+            {
+                return false;
+            }
+
+            if (SpecialMoveType != other.SpecialMoveType)
             {
                 return false;
             }
@@ -385,54 +323,281 @@ namespace Ragnarok.Shogi
         }
 
         /// <summary>
-        /// ハッシュコードを取得します。
+        /// オブジェクトを比較します。
+        /// </summary>
+        public static bool operator ==(Move x, Move y)
+        {
+            return Util.GenericEquals(x, y);
+        }
+
+        /// <summary>
+        /// オブジェクトを比較します。
+        /// </summary>
+        public static bool operator !=(Move x, Move y)
+        {
+            return !(x == y);
+        }
+
+        /// <summary>
+        /// ハッシュ値を計算します。
         /// </summary>
         public override int GetHashCode()
         {
-            // ValueType.GetHashCodeは遅いらしい。。。
-            if (IsSpecialMove)
+            return (
+                BWType.GetHashCode() ^
+                (DstSquare != null ? DstSquare.GetHashCode() : 0) ^
+                (SrcSquare != null ? SrcSquare.GetHashCode() : 0) ^
+                (MovePiece != null ? MovePiece.GetHashCode() : 0) ^
+                DropPieceType.GetHashCode() ^
+                IsPromote.GetHashCode() ^
+                SpecialMoveType.GetHashCode());
+        }
+
+        #region シリアライズ/デシリアライズ
+        [DataMember(Order = 1, IsRequired = true)]
+        private uint serializeBits = 0;
+
+        /// <summary>
+        /// Squareをシリアライズします。
+        /// </summary>
+        /// <remarks>
+        /// 11～99の値にシリアライズされます。
+        /// （0～81の値にシリアライズすると、0の解釈がnull or 11のどちらでも
+        /// 可能になってしまうため、値を減らすことはしていません）
+        /// </remarks>
+        private static byte SerializeSquare(Square square)
+        {
+            if (square == null)
             {
-                return SpecialMoveType.GetHashCode();
+                return 0;
             }
+
+            // 11 ～ 99
+            return (byte)(
+                square.File * (Board.BoardSize + 1) +
+                square.Rank);
+        }
+
+        /// <summary>
+        /// Squareをデシリアライズします。
+        /// </summary>
+        private static Square DeserializeSquare(uint bits)
+        {
+            if (bits == 0)
+            {
+                return null;
+            }
+
+            var file = (int)bits / (Board.BoardSize + 1);
+            var rank = (int)bits % (Board.BoardSize + 1);
+            return new Square(file, rank);
+        }
+
+        /// <summary>
+        /// シリアライズを行います。
+        /// </summary>
+        [CLSCompliant(false)]
+        public uint Serialize()
+        {
+            uint bits = 0;
+
+            // 2bit
+            bits |= (uint)BWType;
+            // 1bit
+            bits |= ((uint)(IsPromote ? 1 : 0) << 2);
+            // 7bit
+            bits |= ((uint)SerializeSquare(DstSquare) << 3);
+            // 7bit
+            bits |= (ActionType == ActionType.Drop ?
+                ((uint)(DropPieceType + 100) << 10) :
+                ((uint)SerializeSquare(SrcSquare) << 10) );
+            // 1bit
+            bits |= ((uint)(HasSameSquareAsPrev ? 1 : 0) << 17);
+
+            // 5bit
+            if (MovePiece != null)
+            {
+                bits |= ((uint)MovePiece.Serialize() << 18);
+            }
+
+            // 5bit
+            if (TookPiece != null)
+            {
+                bits |= ((uint)TookPiece.Serialize() << 23);
+            }
+
+            // 4bit
+            if (SpecialMoveType != SpecialMoveType.None)
+            {
+                bits |= ((uint)SpecialMoveType << 28);
+            }
+
+            return bits;
+        }
+
+        /// <summary>
+        /// デシリアライズを行います。
+        /// </summary>
+        [CLSCompliant(false)]
+        public void Deserialize(uint bits)
+        {
+            uint tmp;
+
+            // 2bit
+            BWType = (BWType)((bits >> 0) & 0x03);
+            // 1bit
+            IsPromote = (((bits >> 2) & 0x01) != 0);
+            // 7bit
+            DstSquare = DeserializeSquare((bits >> 3) & 0x7f);
+            // 7bit
+            tmp = ((bits >> 10) & 0x7f);
+            if (tmp >= 100)
+                DropPieceType = (PieceType)(tmp - 100);
             else
-            {
-                var baseHashCode =
-                    BWType.GetHashCode() ^
-                    SameAsOld.GetHashCode() ^
-                    RelFileType.GetHashCode() ^
-                    RankMoveType.GetHashCode() ^
-                    ActionType.GetHashCode() ^
-                    (Piece != null ? Piece.GetHashCode() : 0) ^
-                    (SrcSquare != null ? SrcSquare.GetHashCode() : 0);
+                SrcSquare = DeserializeSquare(tmp);
+            // 1bit
+            HasSameSquareAsPrev = (((bits >> 17) & 0x01) != 0);
 
-                if (SameAsOld)
-                {
-                    return baseHashCode;
-                }
-                else
-                {
-                    return (
-                        baseHashCode ^
-                        File.GetHashCode() ^
-                        Rank.GetHashCode());
-                }
+            // 5bit
+            tmp = ((bits >> 18) & 0x1f);
+            if (tmp != 0)
+            {
+                MovePiece = new Piece();
+                MovePiece.Deserialize(tmp);
+            }
+
+            // 5bit
+            tmp = ((bits >> 23) & 0x1f);
+            if (tmp != 0)
+            {
+                TookPiece = new Piece();
+                TookPiece.Deserialize(tmp);
+            }
+
+            // 4bit
+            tmp = ((bits >> 28) & 0x0f);
+            if (tmp != 0)
+            {
+                SpecialMoveType = (SpecialMoveType)tmp;
             }
         }
 
         /// <summary>
-        /// == 演算子を実装します。
+        /// シリアライズ前に呼ばれます。
         /// </summary>
-        public static bool operator ==(Move lhs, Move rhs)
+        [OnSerializing()]
+        private void BeforeSerialize(StreamingContext context)
         {
-            return Util.GenericEquals(lhs, rhs);
+            this.serializeBits = Serialize();
         }
 
         /// <summary>
-        /// != 演算子を実装します。
+        /// デシリアライズ後に呼ばれます。
         /// </summary>
-        public static bool operator !=(Move lhs, Move rhs)
+        [OnDeserialized()]
+        private void AfterDeserialize(StreamingContext context)
         {
-            return !(lhs == rhs);
+            Deserialize(this.serializeBits);
+        }
+        #endregion
+
+        /// <summary>
+        /// コンストラクタ
+        /// </summary>
+        public Move()
+        {
+            BWType = BWType.None;
+            IsPromote = false;
+            DropPieceType = PieceType.None;
+        }
+
+        /// <summary>
+        /// 特殊な指し手を生成します。
+        /// </summary>
+        public static Move CreateSpecialMove(BWType turn, SpecialMoveType smoveType)
+        {
+            if (smoveType == SpecialMoveType.None)
+            {
+                throw new ArgumentException("smoveType");
+            }
+
+            return new Move
+            {
+                BWType = turn,
+                SpecialMoveType = smoveType,
+            };
+        }
+
+        /// <summary>
+        /// 移動手を生成します。
+        /// </summary>
+        public static Move CreateMove(BWType turn, Square src, Square dst,
+                                           Piece movePiece, bool isPromote,
+                                           Piece tookPiece = null)
+        {
+            if (turn == BWType.None)
+            {
+                throw new ArgumentException("turn");
+            }
+
+            if (src == null || !src.Validate())
+            {
+                throw new ArgumentException("src");
+            }
+
+            if (dst == null || !dst.Validate())
+            {
+                throw new ArgumentException("dst");
+            }
+
+            if (movePiece == null || !movePiece.Validate())
+            {
+                throw new ArgumentException("movePiece");
+            }
+
+            if (tookPiece != null && !tookPiece.Validate())
+            {
+                throw new ArgumentException("tookPiece");
+            }
+
+            return new Move
+            {
+                BWType = turn,
+                SrcSquare = src,
+                DstSquare = dst,
+                MovePiece = movePiece,
+                IsPromote = isPromote,
+                TookPiece = tookPiece,
+            };
+        }
+
+        /// <summary>
+        /// 駒を打つ手を生成します。
+        /// </summary>
+        public static Move CreateDrop(BWType turn, Square dst,
+                                           PieceType dropPieceType)
+        {
+            if (turn == BWType.None)
+            {
+                throw new ArgumentException("turn");
+            }
+
+            if (dst == null || !dst.Validate())
+            {
+                throw new ArgumentException("dst");
+            }
+
+            if (dropPieceType == PieceType.None)
+            {
+                throw new ArgumentException("dropPieceType");
+            }
+
+            return new Move
+            {
+                BWType = turn,
+                DstSquare = dst,
+                DropPieceType = dropPieceType,
+            };
         }
     }
 }
