+++ Ragnarok.Shogi/ShogiParser.cs
@@ -390,7 +390,7 @@ namespace Ragnarok.Shogi
         /// <summary>
         /// 指し手のパースを行います。
         /// </summary>
-        public static Move ParseMove(string text, bool isNeedEnd)
+        public static LiteralMove ParseMove(string text, bool isNeedEnd)
         {
             var tmp = string.Empty;
 
@@ -400,7 +400,7 @@ namespace Ragnarok.Shogi
         /// <summary>
         /// 差し手のパースを行います。
         /// </summary>
-        public static Move ParseMoveEx(string text, bool isNeedEnd,
+        public static LiteralMove ParseMoveEx(string text, bool isNeedEnd,
                                        ref string parsedText)
         {
             if (string.IsNullOrEmpty(text))
@@ -408,7 +408,7 @@ namespace Ragnarok.Shogi
                 return null;
             }
 
-            var move = new Move()
+            var move = new LiteralMove()
             {
                 OriginalText = text.Trim(),
             };
